feat: add compact link card
