+++ ui/components/LinkCard/index.ts
@@ -1,0 +1,3 @@
+import LinkCard from "./LinkCard";
+
+export default LinkCard;
