+++ ui/sections/Hero/CompactHero.tsx
@@ -19,7 +19,22 @@ export default function CompactHero() {
         },
       }}
     >
-      <Title order={1} fz={42}>
+      <Title fz={32}>
+        The{" "}
+        <Text
+          component="span"
+          variant="gradient"
+          gradient={{
+            from: theme.colors.brand[4],
+            to: theme.colors.secondary[4],
+          }}
+          inherit
+        >
+          one hub
+        </Text>{" "}
+        for
+      </Title>
+      <Title order={1} fz={48}>
         <Text
           component="span"
           variant="gradient"
@@ -31,20 +46,13 @@ export default function CompactHero() {
         >
           {hubName}
         </Text>
-        &apos;s{" "}
-        <Text
-          component="span"
-          variant="gradient"
-          gradient={{
-            from: theme.colors.brand[4],
-            to: theme.colors.secondary[4],
-          }}
-          inherit
-        >
-          Hub
-        </Text>
         .
       </Title>
+      <Text mt="xl" color="dimmed" fz="lg">
+        Tired of keeping track of new websites? Tired of having to update your
+        bookmarks every few weeks? Access all sites from this one page.
+        Everything is up to date. No need to clutter your life anymore!
+      </Text>
     </Container>
   );
 }
